# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "ColumnOut": {
        "additionalProperties": false,
        "description": "docs/07: \"the 34 selectable result columns\" — docs/01 §4 in fact enumerates 36.",
        "properties": {
          "is_factor": {
            "title": "Is Factor",
            "type": "boolean"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "key",
          "label",
          "unit",
          "is_factor"
        ],
        "title": "ColumnOut",
        "type": "object"
      },
...
      "FactorOut": {
        "additionalProperties": false,
        "description": "docs/07: \"factor registry (key, label, family, unit, higher_is_better)\".",
        "properties": {
          "family": {
            "title": "Family",
            "type": "string"
          },
          "higher_is_better": {
            "title": "Higher Is Better",
            "type": "boolean"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "unit": {
            "title": "Unit",
            "type": "string"
          }
        },
        "required": [
          "key",
          "label",
          "family",
          "unit",
          "higher_is_better"
        ],
        "title": "FactorOut",
        "type": "object"
      },
...
      "PipelineRunOut": {
        "additionalProperties": false,
        "properties": {
          "data_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Version"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "trade_date": {
            "format": "date",
            "title": "Trade Date",
            "type": "string"
          }
        },
        "required": [
          "trade_date",
          "status",
          "started_at",
          "finished_at",
          "data_version"
        ],
        "title": "PipelineRunOut",
        "type": "object"
      },
...
      "ProblemOut": {
        "additionalProperties": true,
        "description": "RFC 9457. ``extra=\"allow\"`` carries the per-type members docs/07 attaches.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Instance"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail"
        ],
        "title": "ProblemOut",
        "type": "object"
      },
...
      "StatusOut": {
        "additionalProperties": false,
        "description": "docs/07: \"{ as_of, data_version, last_pipeline_run }\".",
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "data_start_date": {
            "format": "date",
            "title": "Data Start Date",
            "type": "string"
          },
          "data_version": {
            "title": "Data Version",
            "type": "integer"
          },
          "degraded": {
            "title": "Degraded",
            "type": "boolean"
          },
          "last_pipeline_run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PipelineRunOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "as_of",
          "data_version",
          "last_pipeline_run",
          "degraded",
          "data_start_date"
        ],
        "title": "StatusOut",
        "type": "object"
      },
...
      "TradingDaysOut": {
        "additionalProperties": false,
        "description": "docs/07: \"list of trading dates (for the historical date picker)\".",
        "properties": {
          "dates": {
            "items": {
              "format": "date",
              "type": "string"
            },
            "title": "Dates",
            "type": "array"
          },
          "from": {
            "format": "date",
            "title": "From",
            "type": "string"
          },
          "to": {
            "format": "date",
            "title": "To",
            "type": "string"
          }
        },
        "required": [
          "from",
          "to",
          "dates"
        ],
        "title": "TradingDaysOut",
        "type": "object"
      },
      "UniverseOut": {
        "additionalProperties": false,
        "description": "docs/07: \"index_def rows where is_universe\".",
        "properties": {
          "index_id": {
            "title": "Index Id",
            "type": "integer"
          },
          "market_health": {
            "title": "Market Health",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "sort_order": {
            "title": "Sort Order",
            "type": "integer"
          }
        },
        "required": [
          "index_id",
          "slug",
          "name",
          "sort_order",
          "market_health"
        ],
        "title": "UniverseOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "India-equities momentum screener. See docs/07-api-spec.md.",
    "title": "Decile API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/meta/columns": {
      "get": {
        "description": "docs/07: \"the 34 selectable result columns\" — docs/01 §4 in fact enumerates 36.",
        "operationId": "getColumns",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ColumnOut"
                  },
                  "title": "Response Getcolumns",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Invalid screen definition"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Authentication required"
          },
          "402": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Your plan does not include this feature"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Data version is no longer current"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "No trading day available for that date"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Too many requests"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Internal server error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Data pipeline is degraded"
          }
        },
        "summary": "Selectable result columns",
        "tags": [
          "meta"
        ]
      }
    },
    "/api/v1/meta/factors": {
      "get": {
        "description": "docs/07: \"factor registry (key, label, family, unit, higher_is_better)\".\n\nAll 64 of them. docs/01 §3 is headed \"62 ranking factors\" and enumerates 64; see the note in\n``decile_core.factor_registry``.",
        "operationId": "getFactors",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FactorOut"
                  },
                  "title": "Response Getfactors",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Invalid screen definition"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Authentication required"
          },
          "402": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Your plan does not include this feature"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Data version is no longer current"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "No trading day available for that date"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Too many requests"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Internal server error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Data pipeline is degraded"
          }
        },
        "summary": "The factor registry",
        "tags": [
          "meta"
        ]
      }
    },
    "/api/v1/meta/status": {
      "get": {
        "description": "docs/07: \"{ as_of, data_version, last_pipeline_run }\".\n\n``degraded`` is the extra member docs/11 §Reliability implies: \"if the pipeline fails, serve\nthe last good `data_version` with a banner\". The banner needs something to read, and this is\nit — the analytics endpoints keep answering from the last published version.",
        "operationId": "getStatus",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Invalid screen definition"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Authentication required"
          },
          "402": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Your plan does not include this feature"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Data version is no longer current"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "No trading day available for that date"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Too many requests"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Internal server error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Data pipeline is degraded"
          }
        },
        "summary": "Data freshness",
        "tags": [
          "meta"
        ]
      }
    },
    "/api/v1/meta/trading-days": {
      "get": {
        "description": "docs/07: \"list of trading dates (for the historical date picker)\".",
        "operationId": "getTradingDays",
        "parameters": [
          {
            "description": "Inclusive start date.",
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "description": "Inclusive start date.",
              "format": "date",
              "title": "From",
              "type": "string"
            }
          },
          {
            "description": "Inclusive end date.",
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "description": "Inclusive end date.",
              "format": "date",
              "title": "To",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TradingDaysOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Invalid screen definition"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Authentication required"
          },
          "402": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Your plan does not include this feature"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Data version is no longer current"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "No trading day available for that date"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Too many requests"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Internal server error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Data pipeline is degraded"
          }
        },
        "summary": "NSE trading days in a range",
        "tags": [
          "meta"
        ]
      }
    },
    "/api/v1/meta/universes": {
      "get": {
        "description": "docs/07: \"index_def rows where is_universe\" — the fourteen of docs/01 §2.1, in UI order.\n\nServed from ``decile_core.universes`` rather than from ``index_def``: the ids are load-bearing\n(they are ``factor_daily.universe_mask`` bit positions), the seeder writes the table *from*\nthis tuple, and a dashboard-only index must never appear in the screener's dropdown.",
        "operationId": "getUniverses",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UniverseOut"
                  },
                  "title": "Response Getuniverses",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Invalid screen definition"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Authentication required"
          },
          "402": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Your plan does not include this feature"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Data version is no longer current"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "No trading day available for that date"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Too many requests"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Internal server error"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemOut"
                }
              }
            },
            "description": "Data pipeline is degraded"
          }
        },
        "summary": "The selectable universes",
        "tags": [
          "meta"
        ]
      }
    },
...
  }
}